+++ src/core/model/animation/keyframe_transition.hpp
@@ -39,6 +39,9 @@ public:
     };
     Q_ENUM(Descriptive)
 
+    KeyframeTransition() = default;
+    KeyframeTransition(const QPointF& before_handle, const QPointF& after_handle);
+
     const math::bezier::CubicBezierSolver<QPointF>& bezier() const { return bezier_; }
     bool hold() const { return hold_; }